F401 -- Fixed or suppressed linter complaints about unused imports while the public API is still WIP from (- WIP #316 -)
Also includes other minor improvements like normalizing the module path (e.g., pip-licenses -> piplicenses)
