--- pip-licenses/__main__.py
@@ -1,74 +1,0 @@
-#!/usr/bin/env python
-# vim:fenc=utf-8 ff=unix ft=python ts=4 sw=4 sts=4 si et
-"""
-pip-licenses.__main__
-
-MIT License
-
-Copyright (c) 2018-2025 raimon
-Copyright (c) 2025-2026 Mr. Walls
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-"""
-
-from . import (
-    __pkgname__,
-    __version__,
-    __summary__,
-    annotations,
-)
-
-from .cli import (
-    CustomHelpFormatter,
-    FromArg,  # may be deprecated in v6.1? (we can just use argparse or dataclass)
-    OrderArg,  # may be deprecated in v6.1? (we can just use argparse or dataclass)
-    FormatArg,  # may be deprecated in v6.1? (we can just use argparse or dataclass)
-    CustomNamespace,  # TO-BE deprecated in v6.1 (we can just use argparse/dataclass or even an actual namespace)
-    CompatibleArgumentParser, # Preferred since v6.0
-    SelectAction,
-    create_parser,
-)
-
-from .output import (
-    save_if_needs,
-    output_colored,
-    create_output_string,
-)
-
-# placeholder for other main entry point stuff E.g.,
-# __module__ = "pip-licenses"
-
-
-def main() -> None:  # pragma: no cover
-    parser = create_parser()
-    args = parser.parse_args()
-
-    output_string = create_output_string(args)
-
-    output_file = args.output_file
-    save_if_needs(output_file, output_string)
-
-    print(output_string)
-    warn_string = create_warn_string(args)
-    if warn_string:
-        print(warn_string, file=sys.stderr)
-
-
-if __name__ == "__main__":  # pragma: no cover
-    main()
