+++ piplicenses/output/PlainVerticalTable.py
@@ -29,9 +29,9 @@ SOFTWARE.
 
 
 from .. import (
-    __pkgname__,
-    __version__,
-    annotations,
+    __pkgname__,  # noqa: F401 -- Re-export as part of data API
+    __version__,  # noqa: F401 -- Re-export as part of data API
+    annotations,  # noqa: F401 -- from piplicenses.__future__ -> piplicenses.output
 )
