+++ piplicenses/sorting.py
@@ -27,10 +27,10 @@ SOFTWARE.
 """
 
 from . import (
-    __pkgname__,
-    __version__,
-    TYPE_CHECKING,
-    annotations,
+    __pkgname__,  # noqa: F401 -- Re-export as part of data API
+    __version__,  # noqa: F401 -- Re-export as part of data API
+    annotations,  # noqa: F401 -- from piplicenses.__future__
+    TYPE_CHECKING,  # noqa: F401 -- Re-export as part of our internal typing API
 )
 
 from typing import Union
@@ -95,8 +95,8 @@ def case_insensitive_set_diff(
 
 
 __all__ = [
-    "case_insensitive_set_intersect",
-    "case_insensitive_partial_match_set_intersect",
     "case_insensitive_partial_match_set_diff",
+    "case_insensitive_partial_match_set_intersect",
     "case_insensitive_set_diff",
+    "case_insensitive_set_intersect",
 ]
